@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
 	<RuntimeIdentifiers>win10-x64;osx-x64;linux-x64</RuntimeIdentifiers>
@@ -24,31 +24,14 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Content Remove="wwwroot\favicon.ico" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Include="wwwroot\favicon.ico" />
-  </ItemGroup>
-
-  <ItemGroup>
     <PackageReference Include="Flurl" Version="3.0.7" />
     <PackageReference Include="Flurl.Http" Version="3.2.4" />
-    <PackageReference Include="Havit.Blazor.Components.Web.Bootstrap" Version="4.0.4" />
-	<PackageReference Include="Microsoft.Extensions.Logging" Version="7.0.0" />
-	<PackageReference Include="prometheus-net" Version="8.0.1" />
-	<PackageReference Include="prometheus-net.DotNetRuntime" Version="4.4.0" />
-	<PackageReference Include="Serilog.AspNetCore" Version="7.0.0" />
-	<PackageReference Include="Serilog.Enrichers.Span" Version="3.1.0" />
-	<PackageReference Include="Serilog.Settings.Configuration" Version="7.0.0" />
-	<PackageReference Include="Serilog.Sinks.Console" Version="4.1.0" />
-	<PackageReference Include="Serilog.Sinks.File" Version="5.0.0" />
-	<PackageReference Include="Serilog.Sinks.Grafana.Loki" Version="8.1.0" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Api\Api.csproj" />
+    <ProjectReference Include="..\Api.Contract\Api.Contract.csproj" />
     <ProjectReference Include="..\Common\Common.csproj" />
+    <ProjectReference Include="..\SharedUI\SharedUI.csproj" />
   </ItemGroup>
 
   <ItemGroup>
