@@ -1,10 +1,9 @@
-﻿using Common.Database;
-using Common.Dto.Api;
-using Common.Helpers;
+﻿using Api.Contract;
+using Api.Service.Validators;
+using Common.Database;
 using Common.Service;
 using Microsoft.AspNetCore.Mvc;
 using Sync;
-using ErrorResponse = Common.Dto.Api.ErrorResponse;
 
 namespace Api.Controllers;
 
@@ -36,15 +35,16 @@ public class SyncController : Controller
 	[HttpPost]
 	[ProducesResponseType(typeof(SyncPostResponse), StatusCodes.Status201Created)]
 	[ProducesResponseType(typeof(SyncPostResponse), StatusCodes.Status200OK)]
-	[ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status400BadRequest)]
-	[ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status401Unauthorized)]
-	[ProducesResponseType(typeof(ErrorResponse), StatusCodes.Status500InternalServerError)]
+	[ProducesResponseType(typeof(Contract.ErrorResponse), StatusCodes.Status400BadRequest)]
+	[ProducesResponseType(typeof(Contract.ErrorResponse), StatusCodes.Status401Unauthorized)]
+	[ProducesResponseType(typeof(Contract.ErrorResponse), StatusCodes.Status500InternalServerError)]
 	public async Task<ActionResult<SyncPostResponse>> SyncAsync([FromBody] SyncPostRequest request)
 	{
 		var settings = await _settingsService.GetSettingsAsync();
+		var auth = _settingsService.GetGarminAuthentication(settings.Garmin.Email);
 
-		var (isValid, result) = await IsValidAsync(request);
-		if (!isValid) return result;
+		var (isValid, result) = request.IsValidHttp(settings, auth);
+		if (!isValid) return result!;
 
 		SyncResult syncResult = new();
 		try
@@ -53,7 +53,7 @@ public class SyncController : Controller
 		}
 		catch (Exception e)
 		{
-			return StatusCode(StatusCodes.Status500InternalServerError, new ErrorResponse($"Unexpected error occurred: {e.Message}"));
+			return StatusCode(StatusCodes.Status500InternalServerError, new Contract.ErrorResponse($"Unexpected error occurred: {e.Message}"));
 		}
 
 		var response = new SyncPostResponse()
@@ -62,7 +62,7 @@ public class SyncController : Controller
 			PelotonDownloadSuccess = syncResult.PelotonDownloadSuccess,
 			ConverToFitSuccess = syncResult.ConversionSuccess,
 			UploadToGarminSuccess = syncResult.UploadToGarminSuccess,
-			Errors = syncResult.Errors.Select(e => new ErrorResponse(e.Message)).ToList()
+			Errors = syncResult.Errors.Select(e => new Contract.ErrorResponse(e.Message)).ToList()
 		};
 
 		if (!response.SyncSuccess)
@@ -98,29 +98,5 @@ public class SyncController : Controller
 		};
 
 		return response;
-	}
-
-	async Task<(bool, ActionResult)> IsValidAsync(SyncPostRequest request)
-	{
-		ActionResult result = new OkResult();
-
-		if (request.CheckIsNull("PostRequest", out result))
-			return (false, result);
-
-		var settings = await _settingsService.GetSettingsAsync();
-		if (settings.Garmin.Upload && settings.Garmin.TwoStepVerificationEnabled)
-		{
-			var auth = _settingsService.GetGarminAuthentication(settings.Garmin.Email);
-			if (auth is null || !auth.IsValid(settings))
-			{
-				result = new UnauthorizedObjectResult(new ErrorResponse("Must initialize Garmin two factor auth token before sync can be preformed.", ErrorCode.NeedToInitGarminMFAAuth));
-				return (false, result);
-			}
-		}
-
-		if (request.WorkoutIds.CheckDoesNotHaveAny(nameof(request.WorkoutIds), out result))
-			return (false, result);
-
-		return (true, result);
 	}
 }
