@@ -1,23 +1,11 @@
 using Api.Services;
 using Common;
 using Common.Database;
-using Common.Http;
-using Common.Observe;
-using Common.Service;
 using Common.Stateful;
-using Conversion;
-using Garmin;
-using Garmin.Auth;
-using Microsoft.Extensions.Caching.Memory;
-using Peloton;
-using Peloton.AnnualChallenge;
-using Philosowaffle.Capability.ReleaseChecks;
 using Prometheus;
 using Serilog;
-using Serilog.Enrichers.Span;
 using Serilog.Events;
-using Serilog.Settings.Configuration;
-using Sync;
+using SharedStartup;
 using System.Reflection;
 
 ///////////////////////////////////////////////////////////
@@ -31,8 +19,7 @@ Statics.ConfigPath = Path.Join(Environment.CurrentDirectory, "configuration.loca
 ///////////////////////////////////////////////////////////
 /// HOST
 ///////////////////////////////////////////////////////////
-var builder = WebApplication
-				.CreateBuilder(args);
+var builder = WebApplication.CreateBuilder(args);
 
 var configProvider = builder.Configuration.AddJsonFile(Statics.ConfigPath, optional: true, reloadOnChange: true)
 				.AddEnvironmentVariables(prefix: "P2G_")
@@ -43,19 +30,11 @@ ConfigurationSetup.LoadConfigValues(builder.Configuration, config);
 
 builder.WebHost.UseUrls(config.Api.HostUrl);
 
-builder.Host.UseSerilog((ctx, logConfig) =>
-{
-	logConfig
-	.ReadFrom.Configuration(ctx.Configuration, new ConfigurationReaderOptions() { SectionName = $"{nameof(Observability)}:Serilog" })
-	.Enrich.WithSpan()
-	.Enrich.FromLogContext();
-});
-
-builder.Host.ConfigureServices(services => services.AddHostedService<BackgroundSyncJob>());
-
 ///////////////////////////////////////////////////////////
 /// SERVICES
 ///////////////////////////////////////////////////////////
+
+builder.Services.AddHostedService<BackgroundSyncJob>();
 
 builder.Services.AddControllers();
 builder.Services.AddEndpointsApiExplorer();
@@ -72,53 +51,9 @@ builder.Services.AddSwaggerGen(c =>
 		c.IncludeXmlComments(docPath);
 });
 
-// CACHE
-builder.Services.AddSingleton<IMemoryCache, MemoryCache>();
+builder.Services.ConfigureP2GApiServices();
 
-// CONVERT
-builder.Services.AddSingleton<IConverter, FitConverter>();
-builder.Services.AddSingleton<IConverter, TcxConverter>();
-builder.Services.AddSingleton<IConverter, JsonConverter>();
-
-// GARMIN
-builder.Services.AddSingleton<IGarminAuthenticationService, GarminAuthenticationService>();
-builder.Services.AddSingleton<IGarminUploader, GarminUploader>();
-builder.Services.AddSingleton<IGarminApiClient, Garmin.ApiClient>();
-
-// IO
-builder.Services.AddSingleton<IFileHandling, IOWrapper>();
-
-// MIGRATIONS
-builder.Services.AddSingleton<IDbMigrations, DbMigrations>();
-
-// PELOTON
-builder.Services.AddSingleton<IPelotonApi, Peloton.ApiClient>();
-builder.Services.AddSingleton<IPelotonService, PelotonService>();
-builder.Services.AddSingleton<IAnnualChallengeService, AnnualChallengeService>();
-
-// RELEASE CHECKS
-builder.Services.AddGitHubReleaseChecker();
-
-// SETTINGS
-builder.Services.AddSingleton<ISettingsDb, SettingsDb>();
-builder.Services.AddSingleton<ISettingsService, SettingsService>();
-
-// SYNC
-builder.Services.AddSingleton<ISyncStatusDb, SyncStatusDb>();
-builder.Services.AddSingleton<ISyncService, SyncService>();
-
-// USERS
-builder.Services.AddSingleton<IUsersDb, UsersDb>();
-
-FlurlConfiguration.Configure(config.Observability);
-Tracing.EnableApiTracing(builder.Services, config.Observability.Jaeger);
-
-Log.Logger = new LoggerConfiguration()
-				.ReadFrom.Configuration(builder.Configuration, new ConfigurationReaderOptions() { SectionName = $"{nameof(Observability)}:Serilog" })
-				.Enrich.FromLogContext()
-				.CreateLogger();
-
-Logging.LogSystemInformation();
+ObservabilityStartup.ConfigureApi(builder.Services, builder.Configuration, config);
 Common.Observe.Metrics.CreateAppInfo();
 
 ///////////////////////////////////////////////////////////
@@ -127,8 +62,17 @@ Common.Observe.Metrics.CreateAppInfo();
 
 var app = builder.Build();
 
-// Setup initial Tracing Source
-Tracing.Source = new(Statics.TracingService);
+if (Log.IsEnabled(LogEventLevel.Verbose))
+	app.UseSerilogRequestLogging();
+
+if (config.Observability.Prometheus.Enabled)
+{
+	Log.Information("Metrics Enabled");
+	Common.Observe.Metrics.EnableCollector(config.Observability.Prometheus);
+
+	app.MapMetrics();
+	app.UseHttpMetrics();
+}
 
 app.UseCors(options =>
 {
@@ -140,22 +84,11 @@ app.UseCors(options =>
 app.UseSwagger();
 app.UseSwaggerUI();
 
-if (Log.IsEnabled(LogEventLevel.Verbose))
-	app.UseSerilogRequestLogging();
-
 app.Use((context, next) =>
 {
 	return next.Invoke();
 });
 
-if (config.Observability.Prometheus.Enabled)
-{
-	Log.Information("Metrics Enabled");
-	Common.Observe.Metrics.EnableCollector(config.Observability.Prometheus);
-
-	app.MapMetrics();
-	app.UseHttpMetrics();
-}
 
 //app.UseHttpsRedirection();
 app.UseAuthorization();
