@@ -1,9 +1,5 @@
 ﻿using Common.Dto;
 using Microsoft.Extensions.Configuration;
-using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.IO;
 
 namespace Common;
 
@@ -13,8 +9,8 @@ public static class ConfigurationSetup
 	{
 		provider.GetSection(nameof(App)).Bind(config.App);
 		provider.GetSection(nameof(Format)).Bind(config.Format);
-		provider.GetSection(nameof(Peloton)).Bind(config.Peloton);
-		provider.GetSection(nameof(Garmin)).Bind(config.Garmin);
+		provider.GetSection("Peloton").Bind(config.Peloton);
+		provider.GetSection("Garmin").Bind(config.Garmin);
 	}
 
 	public static void LoadConfigValues(IConfiguration provider, AppConfiguration config)
@@ -45,170 +41,11 @@ public class AppConfiguration
 	public Developer Developer { get; set; }
 }
 
-/// <summary>
-/// Settings that can be looked up after app start, changed on demand, and saved to the SettingsDb.
-/// </summary>
-public class Settings
-{
-	public Settings()
-	{
-		App = new App();
-		Format = new Format();
-		Peloton = new Peloton();
-		Garmin = new Garmin();
-	}
-
-	public App App { get; set; }
-	public Format Format { get; set; }
-	public Peloton Peloton { get; set; }
-	public Garmin Garmin { get; set; }
-}
-
-public class App
-{
-	public App()
-	{
-		OutputDirectory = Path.Join(Environment.CurrentDirectory, "output");
-		WorkingDirectory = Path.Join(Environment.CurrentDirectory, "working");
-
-		CheckForUpdates = true;
-		EnablePolling = false;
-		PollingIntervalSeconds = 86400; // 1 day
-	}
-
-	[DisplayName("Output Directory")]
-	[Description("Where downloaded and converted files should be saved to.")] 
-	public string OutputDirectory { get; set; }
-	[DisplayName("Working Directory")]
-	[Description("The directory where P2G can work. When running, P2G will create and delete files and needs a dedicated directory to do that.")]
-	public string WorkingDirectory { get; set; }
-
-	[DisplayName("Enable Polling")]
-	[Description("Enabled if you wish P2G to run continuously and poll Peloton for new workouts.")]
-	public bool EnablePolling { get; set; }
-	[DisplayName("Polling Interval in Seconds")]
-	[Description("The polling interval in seconds determines how frequently P2G should check for new workouts. Be warned, that setting this to a frequency of hourly or less may get you flagged by Peloton as a bad actor and they may reset your password.")]
-	public int PollingIntervalSeconds { get; set; }
-	[Obsolete]
-	public bool? PythonAndGUploadInstalled { get; set; }
-	public bool CloseWindowOnFinish { get; set; }
-	public bool CheckForUpdates { get; set; }
-
-
-	public static string DataDirectory = Path.GetFullPath(Path.Join(Environment.CurrentDirectory, "data"));
-	public string FailedDirectory => Path.GetFullPath(Path.Join(OutputDirectory, "failed"));
-	public string DownloadDirectory => Path.GetFullPath(Path.Join(WorkingDirectory, "downloaded"));
-	public string UploadDirectory => Path.GetFullPath(Path.Join(WorkingDirectory, "upload"));
-
-}
-
-public class Format
-{
-	public Format()
-	{
-		Cycling = new Cycling();
-		Running = new Running();
-		Rowing = new Rowing();
-		Strength= new Strength();
-	}
-
-	[DisplayName("FIT")]
-	[Description("Enabled indicates you wish downloaded workouts to be converted to FIT")]
-	public bool Fit { get; set; }
-	[DisplayName("JSON")]
-	[Description("Enabled indicates you wish downloaded workouts to be converted to JSON")]
-	public bool Json { get; set; }
-	[DisplayName("TCX")]
-	[Description("Enabled indicates you wish downloaded workouts to be converted to TCX.")]
-	public bool Tcx { get; set; }
-	[DisplayName("Save a local copy")]
-	[Description("Save any converted workouts to your specified Output Directory")]
-	public bool SaveLocalCopy { get; set; }
-	[DisplayName("Include Time in HR Zones")]
-	[Description("Only use this if you are unable to configure your Max HR on Garmin Connect. When set to True, P2G will attempt to capture the time spent in each HR Zone per the data returned by Peloton.")]
-	public bool IncludeTimeInHRZones { get; set; }
-	[DisplayName("Include Time in Power Zones")]
-	[Description("Only use this if you are unable to configure your FTP and Power Zones on Garmin Connect. When set to True, P2G will attempt to capture the time spent in each Power Zone per the data returned by Peloton.")]
-	public bool IncludeTimeInPowerZones { get; set; }
-	[DisplayName("Device Info Path")]
-	[Description("The path to your deviceInfo.xml file.")]
-	public string DeviceInfoPath { get; set; }
-	public Cycling Cycling { get; set; }
-	public Running Running { get; set; }
-	public Rowing Rowing { get; init; }
-	public Strength Strength { get; init; }
-}
-
-public record Cycling
-{
-	public PreferredLapType PreferredLapType { get; set; }
-}
-
-public record Running
-{
-	public PreferredLapType PreferredLapType { get; set; }
-}
-
-public record Rowing
-{
-	public PreferredLapType PreferredLapType { get; set; }
-}
-
-public record Strength
-{
-	/// <summary>
-	/// When no Rep information is provided by Peloton, P2G will calculate number
-	/// of reps based on this default value. Example, if your DefaultNumSecondsPerRep is 3,
-	/// and the Exercise duration was 15 seconds, then P2G would credit you with 5 reps for that
-	/// exercise.
-	/// </summary>
-	public int DefaultSecondsPerRep { get; set; } = 3;
-}
-
-public enum PreferredLapType
-{
-	Default = 0,
-	Distance = 1,
-	Class_Segments = 2,
-	Class_Targets = 3
-}
-
-public class Peloton : ICredentials
-{
-	public Peloton()
-	{
-		ExcludeWorkoutTypes = new List<WorkoutType>();
-		NumWorkoutsToDownload = 5;
-	}
-
-	public EncryptionVersion EncryptionVersion { get; set; }
-	public string Email { get; set; }
-	public string Password { get; set; }
-	public int NumWorkoutsToDownload { get; set; }
-	public ICollection<WorkoutType> ExcludeWorkoutTypes { get; set; }
-}
-
-public class Garmin : ICredentials
-{
-	public Garmin()
-	{
-		UploadStrategy = UploadStrategy.NativeImplV1;
-	}
-
-	public EncryptionVersion EncryptionVersion { get; set; }
-	public string Email { get; set; }
-	public string Password { get; set; }
-	public bool TwoStepVerificationEnabled { get; set; }
-	public bool Upload { get; set; }
-	public FileFormat FormatToUpload { get; set; }
-	public UploadStrategy UploadStrategy { get; set; }
-}
-
 public class ApiSettings
 {
 	public ApiSettings()
 	{
-		HostUrl = "http://localhost";
+		HostUrl = "http://*:8080";
 	}
 
 	public string HostUrl { get; set; }
@@ -218,7 +55,7 @@ public class WebUISettings
 {
 	public WebUISettings()
 	{
-		HostUrl = "http://localhost";
+		HostUrl = "http://*:8080";
 	}
 
 	public string HostUrl { get; set; }
@@ -252,20 +89,6 @@ public class Prometheus
 public class Developer
 {
 	public string UserAgent { get; set; }
-}
-
-public enum UploadStrategy : byte
-{
-	PythonAndGuploadInstalledLocally = 0,
-	WindowsExeBundledPython = 1,
-	NativeImplV1 = 2
-}
-
-public enum FileFormat : byte
-{
-	Fit = 0,
-	Tcx = 1,
-	Json = 2
 }
 
 public enum EncryptionVersion : byte
