@@ -7,6 +7,8 @@
 		public const string ConsoleAppName = "p2g_console";
 		public const string ApiName = "p2g_api";
 		public const string WebUIName = "p2g_webui";
-		public const string AppVersion = "3.6.1";
+		public const string ClientUIName = "p2g_clientui";
+
+		public const string AppVersion = "4.2.0-rc";
 	}
 }
