@@ -1,4 +1,5 @@
 ﻿using Common;
+using Common.Dto;
 using Common.Observe;
 using Common.Service;
 using Common.Stateful;
@@ -6,7 +7,6 @@ using Garmin.Auth;
 using Prometheus;
 using Serilog;
 using System;
-using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
@@ -72,19 +72,8 @@ namespace Garmin
 			using var metrics = WorkoutUploadDuration
 								.WithLabels(files.Count().ToString()).NewTimer();
 
-			switch (settings.Garmin.UploadStrategy)
-			{
-				case UploadStrategy.PythonAndGuploadInstalledLocally:
-				case UploadStrategy.WindowsExeBundledPython:
-					UploadViaPython(files, settings);
-					_logger.Information("Upload complete.");
-					return;
-				case UploadStrategy.NativeImplV1:
-				default:
-					await UploadAsync(files, settings);
-					_logger.Information("Upload complete.");
-					return;
-			}
+			await UploadAsync(files, settings);
+			_logger.Information("Upload complete.");
 		}
 
 		private async Task UploadAsync(string[] files, Settings settings)
@@ -119,84 +108,11 @@ namespace Garmin
 			await Task.Delay(waitDuration);
 		}
 
-		private void UploadViaPython(string[] files, Settings settings)
-		{
-			using var tracing = Tracing.Trace($"{nameof(GarminUploader)}.{nameof(UploadViaPython)}.UploadToGarminViaPython")
-										.WithTag(TagKey.Category, "gupload");
-
-			settings.Garmin.EnsureGarminCredentialsAreProvided();
-
-			ProcessStartInfo start = new ProcessStartInfo();
-			var paths = String.Join(" ", files.Select(p => $"\"{p}\""));
-			var cmd = string.Empty;
-
-			if (settings.Garmin.UploadStrategy == UploadStrategy.PythonAndGuploadInstalledLocally)
-			{
-				start.FileName = "gupload";
-				cmd = $"-u {settings.Garmin.Email} -p {settings.Garmin.Password} {paths}";
-			} else
-			{
-				paths = String.Join(" ", files.Select(f => $"\"{Path.GetFullPath(f)}\""));
-				start.FileName = Path.Join(Environment.CurrentDirectory, "python", "upload.exe");
-				cmd = $"-ge {settings.Garmin.Email} -gp {settings.Garmin.Password} -f {paths}";
-			}
-
-			_logger.Information("Beginning Garmin Upload.");
-			_logger.Information("Uploading to Garmin with the following parameters: {@File} {@Command}", start.FileName, cmd.Replace(settings.Garmin.Email, "**email**").Replace(settings.Garmin.Password, "**password**"));
-
-			start.Arguments = cmd;
-			start.UseShellExecute = false;
-			start.CreateNoWindow = true;
-			start.RedirectStandardOutput = true;
-			start.RedirectStandardError = true;
-
-			FilesToUpload.Set(files.Length);
-			if (files.Length > 20)
-				_logger.Information("Detected large number of files for upload to Garmin. Please be patient, this could take a while.");
-			using var process = Process.Start(start);
-			process.WaitForExit();
-
-			var stderr = process.StandardError.ReadToEnd();
-			var stdout = process.StandardOutput.ReadToEnd();
-
-			if (!string.IsNullOrEmpty(stdout))
-				_logger.Information(stdout);
-
-			// Despite coming from StandardError, this is not necessarily an error, just the output
-			if (!string.IsNullOrEmpty(stderr))
-				_logger.Information("GUpload: {Output}", stderr);
-
-			if (process.HasExited && process.ExitCode != 0)
-			{
-				FailedUploadAttemptsGauge.Inc();
-				throw new GarminUploadException("GUpload returned an error code. Failed to upload workouts.", process.ExitCode);
-			} else
-			{
-				FailedUploadAttemptsGauge.Set(0);
-				FilesToUpload.Set(0);
-			}
-		}
-
 		public static void ValidateConfig(Settings config)
 		{
 			if (config.Garmin.Upload == false) return;
 
 			config.Garmin.EnsureGarminCredentialsAreProvided();
-
-			if (config.App.EnablePolling && config.Garmin.TwoStepVerificationEnabled)
-				throw new ArgumentException("App.EnablePolling cannot be true when Garmin.TwoStepVerificationEnabled is true.");
-
-			if (config.App.PythonAndGUploadInstalled.HasValue)
-			{
-				_logger.Warning("App.PythonAndGuploadInstalledLocally setting is deprecated and will be removed in a future release. Please swith to using Garmin.UploadStrategy config.");
-
-				if (config.Garmin.UploadStrategy == UploadStrategy.PythonAndGuploadInstalledLocally
-					&& config.App.PythonAndGUploadInstalled.Value == false)
-				{
-					config.Garmin.UploadStrategy = UploadStrategy.WindowsExeBundledPython;
-					_logger.Warning("Detected use of deprecated config App.PythonAndGuploadInstalledLocally, setting Garmin.UploadStrategy to WindowsExeBundledPython=1");
-				}
-			}
 		}
 	}
 }
