@@ -1,11 +1,12 @@
-﻿using Flurl.Http;
+﻿using Common.Dto;
+using Flurl.Http;
 using System;
 
 namespace Peloton;
 
 public static class Extensions
 {
-	public static void EnsurePelotonCredentialsAreProvided(this Common.Peloton settings)
+	public static void EnsurePelotonCredentialsAreProvided(this PelotonSettings settings)
 	{
 		if (string.IsNullOrEmpty(settings.Email))
 			throw new ArgumentException("Peloton Email must be set.", nameof(settings.Email));
