@@ -1,5 +1,4 @@
 ﻿using Common.Dto;
-using System;
 using System.Collections.Generic;
 
 namespace Sync;
@@ -8,24 +7,12 @@ public class SyncResult
 {
 	public SyncResult()
 	{
-		Errors = new List<IServiceError>();
+		Errors = new List<ServiceError>();
 	}
 
 	public bool SyncSuccess { get; set; }
 	public bool PelotonDownloadSuccess { get; set; }
 	public bool? ConversionSuccess { get; set; }
 	public bool? UploadToGarminSuccess { get; set; }
-	public ICollection<IServiceError> Errors { get; set; }
-}
-
-public class ErrorResponse : IServiceError
-{
-	public ErrorResponse()
-	{
-		Message = string.Empty;
-	}
-
-	public Exception? Exception { get; init; }
-	public string Message { get; init; }
-	public bool IsServerException { get; init; }
+	public ICollection<ServiceError> Errors { get; set; }
 }
