@@ -3,6 +3,7 @@ using System;
 using Common;
 using FluentAssertions;
 using System.IO;
+using Common.Dto;
 
 namespace UnitTests.Common
 {
@@ -63,7 +64,6 @@ namespace UnitTests.Common
 			config.Garmin.Password.Should().BeNull();
 			config.Garmin.FormatToUpload.Should().Be(FileFormat.Fit);
 			config.Garmin.Upload.Should().BeFalse();
-			config.Garmin.UploadStrategy.Should().Be(UploadStrategy.NativeImplV1);
 		}
 
 		[Test]
