@@ -1,4 +1,5 @@
-﻿using Common.Dto.Peloton;
+﻿using Common.Dto;
+using Common.Dto.Peloton;
 using Common.Helpers;
 using FluentAssertions;
 using NUnit.Framework;
@@ -27,7 +28,26 @@ public class WorkoutHelperTests
 			}
 		};
 
-		return WorkoutHelper.GetTitle(workout);
+		return WorkoutHelper.GetTitle(workout, new Format());
+	}
+
+	[TestCase("My Title", "é", ExpectedResult = "My_Title_with_é")]
+	[TestCase("My Title", "ä", ExpectedResult = "My_Title_with_ä")]
+	public string GetTitle_Should_Handle_SpecialChars(string title, string instructor)
+	{
+		var workout = new Workout()
+		{
+			Ride = new Ride()
+			{
+				Title = title,
+				Instructor = new Instructor()
+				{
+					Name = instructor
+				}
+			}
+		};
+
+		return WorkoutHelper.GetTitle(workout, new Format());
 	}
 
 	[Test]
@@ -38,7 +58,7 @@ public class WorkoutHelperTests
 			Id = "someId"
 		};
 
-		var title = WorkoutHelper.GetTitle(workout);
+		var title = WorkoutHelper.GetTitle(workout, new Format());
 		title.Should().Be("someId");
 	}
 
@@ -53,7 +73,7 @@ public class WorkoutHelperTests
 			}
 		};
 
-		var title = WorkoutHelper.GetTitle(workout);
+		var title = WorkoutHelper.GetTitle(workout, new Format());
 		title.Should().Be("My_Title");
 	}
 
@@ -69,7 +89,44 @@ public class WorkoutHelperTests
 			}
 		};
 
-		var title = WorkoutHelper.GetTitle(workout);
+		var title = WorkoutHelper.GetTitle(workout, new Format());
 		title.Should().Be("My_Title");
+	}
+
+	[Test]
+	public void GetTitle_NullTemplate_ShouldReturn_DefaultTemplate()
+	{
+		var workout = new Workout()
+		{
+			Ride = new Ride()
+			{
+				Title = "My Title",
+				Instructor = new Instructor() { Name = "Instructor"}
+			}
+		};
+
+		var title = WorkoutHelper.GetTitle(workout, new Format());
+		title.Should().Be("My_Title_with_Instructor");
+	}
+
+	[Test]
+	public void GetTitle_With_Template_ShouldReturn_TemplateAppliedToTitle()
+	{
+		var format = new Format()
+		{
+			WorkoutTitleTemplate = "{{PelotonInstructorName}} - {{PelotonWorkoutTitle}}"
+		};
+
+		var workout = new Workout()
+		{
+			Ride = new Ride()
+			{
+				Title = "My Title",
+				Instructor = new Instructor() { Name = "Instructor" }
+			}
+		};
+
+		var title = WorkoutHelper.GetTitle(workout, format);
+		title.Should().Be("Instructor_-_My_Title");
 	}
 }
