@@ -5,4 +5,4 @@
     Layout = "_Layout";
 }
 
-<component type="typeof(App)" render-mode="ServerPrerendered" />
+<component type="typeof(SharedUI.App)" render-mode="ServerPrerendered" />
