@@ -8,10 +8,10 @@
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <base href="~/" />
-    <link rel="stylesheet" href="css/bootstrap/bootstrap.min.css" />
-    <link rel="stylesheet" href="lib/bootstrap-icons/font/bootstrap-icons.css" />
-    <link href="css/site.css" rel="stylesheet" />
-    <link href="WebUI.styles.css" rel="stylesheet" />
+    <link rel="stylesheet" href="_content/SharedUI/css/bootstrap/bootstrap.min.css" />
+    <link rel="stylesheet" href="_content/SharedUI/lib/bootstrap-icons/font/bootstrap-icons.css" />
+    <link rel="stylesheet" href="_content/SharedUI/css/site.css" />
+    <link rel="stylesheet" href="WebUI.styles.css" />
     <component type="typeof(HeadOutlet)" render-mode="ServerPrerendered" />
 </head>
 <body>
@@ -29,8 +29,8 @@
     </div>
 
     <script src="_framework/blazor.server.js"></script>
-    <script src="lib/font-awesome/js/all.js"></script>
-    <script src="lib/popperjs/umd/popper.min.js"></script>
-    <script src="lib/bootstrap/js/bootstrap.min.js"></script>
+    <script src="_content/SharedUI/lib/font-awesome/js/all.js"></script>
+    <script src="_content/SharedUI/lib/popperjs/umd/popper.min.js"></script>
+    <script src="_content/SharedUI/lib/bootstrap/js/bootstrap.min.js"></script>
 </body>
 </html>
